Updated Target to iOS 6 and ARM 64bit
Updated the iOS target to 6 and changed the architecture to support ARM
64 bit as well. Removing depreciated variables and methods. Making sure
that all conversions from long/NSInteger to int are properly cast.
Updated some methods and variables to use NSInteger instead to preserve
the correct accuracy without casting.

+++ src/TapkuLibrary/NSDate+TKCategory.h
@@ -35,15 +35,15 @@
 
 
 struct TKDateInformation {
-	int day;
-	int month;
-	int year;
+	NSInteger day;
+	NSInteger month;
+	NSInteger year;
 	
-	int weekday;
+	NSInteger weekday;
 	
-	int minute;
-	int hour;
-	int second;
+	NSInteger minute;
+	NSInteger hour;
+	NSInteger second;
 	
 };
 typedef struct TKDateInformation TKDateInformation;
@@ -67,7 +67,7 @@ typedef struct TKDateInformation TKDateInformation;
 
 
 - (BOOL) isSameDay:(NSDate*)anotherDate;
-- (int) monthsBetweenDate:(NSDate *)toDate;
+- (NSInteger) monthsBetweenDate:(NSDate *)toDate;
 - (NSInteger) daysBetweenDate:(NSDate*)date;
 
 /** Returns a Boolean value that indicates whether the date object is that same date information as the current day.
